Feat: Add DeckDelegate as a view model of CardDeckView #13
